LPS-191852 Create recuder util file and move functions there
