+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/ModelBuilderContext/objectFolderReducer.ts
@@ -17,76 +17,11 @@ import {
 	TAction,
 	TState,
 } from '../types';
+import {
+	fieldsCustomSort,
+	getNonOverlappingEdges,
+} from './objectFolderReducerUtil';
 import {TYPES} from './typesEnum';
-
-function fieldsCustomSort(objectFields: ObjectFieldNode[]) {
-	const fieldOrder = ['id', 'externalReferenceCode'];
-
-	const compareFields = (a: ObjectFieldNode, b: ObjectFieldNode) => {
-		const aIndex = fieldOrder.indexOf(a.name as string);
-		const bIndex = fieldOrder.indexOf(b.name as string);
-
-		if (aIndex !== -1 && bIndex !== -1) {
-			return aIndex - bIndex;
-		}
-		else if (aIndex !== -1) {
-			return -1;
-		}
-		else if (bIndex !== -1) {
-			return 1;
-		}
-
-		if (a.required && !b.required) {
-			return -1;
-		}
-		else if (!a.required && b.required) {
-			return 1;
-		}
-
-		return 0;
-	};
-
-	return objectFields.sort(compareFields);
-}
-
-function findEdgesWithSameSourceAndTarget(edges: Edge<EdgeData>[]) {
-	const duplicates = edges.filter((edge, index, self) => {
-		const foundIndex = self.findIndex(
-			(item) =>
-				(item.source === edge.source && item.target === edge.target) ||
-				(item.source === edge.target && item.target === edge.source)
-		);
-
-		return foundIndex !== -1 && foundIndex !== index;
-	});
-
-	return duplicates;
-}
-
-function getNonOverlappingEdges(allEdges: Edge<EdgeData>[]) {
-	const overlapEdges = findEdgesWithSameSourceAndTarget(allEdges);
-
-	const nonOverlappingEdges = overlapEdges.map((edge) => {
-		const newEdge = {
-			...edge,
-			data: {
-				...edge.data,
-				sourceY: 50,
-				targetY: 50,
-			},
-		} as Edge<EdgeData>;
-
-		return newEdge;
-	});
-
-	const filteredEdges: Edge<EdgeData>[] = allEdges.filter((edge) => {
-		return !nonOverlappingEdges.some(
-			(nonOverlappingEdge) => edge.id === nonOverlappingEdge.id
-		);
-	});
-
-	return [...filteredEdges, ...nonOverlappingEdges];
-}
 
 export function objectFolderReducer(state: TState, action: TAction) {
 	switch (action.type) {
