+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/ModelBuilderContext/objectFolderReducerUtil.d.ts
@@ -1,0 +1,13 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {Edge} from 'react-flow-renderer';
+import {EdgeData, ObjectFieldNode} from '../types';
+export declare function fieldsCustomSort(
+	objectFields: ObjectFieldNode[]
+): ObjectFieldNode[];
+export declare function getNonOverlappingEdges(
+	allEdges: Edge<EdgeData>[]
+): Edge<EdgeData>[];
